WLA-62: Enable multiple decs in pdf generation
